split pooling tests into separate module
